requirements.txt
groq
langchain 
langchain-groq
groq 
python-dotenv
langchain-huggingface